fixing filters and change password

+++ lib/ui/pages/main_page/settings_page/pages/change_password_page.dart
@@ -74,14 +74,16 @@ class ChangePasswordPage extends StatelessWidget {
                                       if (_formKey.currentState?.validate() ??
                                           false) {
                                         await settingsStore.changePassword();
-                                        SuccessDialog();
-                                        Navigator.of(context,
-                                                rootNavigator: true)
-                                            .pushNamedAndRemoveUntil(
-                                          SignInPage.routeName,
-                                          (route) => false,
-                                        );
-                                        Storage.deleteAllFromSecureStorage();
+                                        if (settingsStore.isSuccess) {
+                                          SuccessDialog();
+                                          Navigator.of(context,
+                                                  rootNavigator: true)
+                                              .pushNamedAndRemoveUntil(
+                                            SignInPage.routeName,
+                                            (route) => false,
+                                          );
+                                          Storage.deleteAllFromSecureStorage();
+                                        }
                                       }
                                     }
                                   : null,
